feat(api): extract Conversation read model from chat.ts
listConversations/unreadCount/listEntries inlined multi-table inbox joins + shaping. Move the query side to chat-read-model.ts; router procedures delegate. Access checks stay in the router and stay split (ADR-0004) — read model has zero access logic. chat.ts 544->262. +4 harness tests.
